django_restapi/requirements.txt
certifi==2021.10.8
Django==2.2.5
djangorestframework==3.12.4
pytz==2021.3
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
wincertstore==0.2
